+++ src/components/Word.tsx
