+++ src/components/WordCounter.tsx
